requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
python-multipart==0.0.19
pydantic==2.10.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
